doc: Remove extra newlines and unindentations.
Using spacing instead of indents is contrary to most LaTeX styles
guides (at least while using the default document classes), and more
importantly, was applied very inconsistently.

Several newlines were used to make pseudo-lists, so I replaced those
with real lists.

Some unindentation was used to make paragraphs appear as continuations
of an earlier paragraph. The correct way to do that is to either not
leave the blank line, or comment out the blank (such as the case with
intervening environments).